Revert "search i egen komponent, skjule contentree funker ikke"
This reverts commit 450eaeb784dab65c782851559df134c938a7cd61.
