+++ xp-archive/client/contentTree/NavigationBar.module.css
@@ -3,3 +3,7 @@
     grid-area: left;
     overflow-y: auto;
 }
+
+.heading {
+    margin: 1rem;
+}
