+++ xp-archive/client/contentTree/NavigationBar.tsx
@@ -1,8 +1,11 @@
-import React from 'react';
-import { Tabs } from '@navikt/ds-react';
+import React, { useState } from 'react';
+import { Heading, Tabs, Search } from '@navikt/ds-react';
 import { LayerPanel } from './layerPanel/LayerPanel';
 import { useAppState } from 'client/context/appState/useAppState';
-import { SearchSection, useSearch } from './search/SearchSection';
+import { fetchJson } from '@common/shared/fetchUtils';
+import { SearchResponse } from 'shared/types';
+import { SearchResult } from './search/SearchResult';
+
 import style from './NavigationBar.module.css';
 
 const locales = ['no', 'en', 'nn', 'se'] as const;
@@ -20,12 +23,65 @@ const getLabel = (locale: Locale) => {
 
 export const NavigationBar = () => {
     const { setSelectedLocale } = useAppState();
-    const { searchResultIsOpen } = useSearch();
+    const [searchResultIsOpen, setSearchResultIsOpen] = useState(false);
+    const [searchQuery, setSearchQuery] = useState('');
+    const [isLoading, setIsLoading] = useState(false);
+    const [searchResult, setSearchResult] = useState<SearchResponse>({
+        hits: [],
+        total: 0,
+        hasMore: false,
+        query: '',
+    });
+
+    const SEARCH_API = `${import.meta.env.VITE_APP_ORIGIN}/xp/api/search`;
+
+    const searchData = async () => {
+        setIsLoading(true);
+        const result = await fetchJson<SearchResponse>(SEARCH_API, {
+            params: { query: searchQuery },
+        });
+        if (result) {
+            setSearchResult(result);
+        }
+        setSearchResultIsOpen(true);
+        setIsLoading(false);
+        return result;
+    };
+
+    const closeSearchResult = () => {
+        setSearchResultIsOpen(false);
+        setSearchQuery('');
+    };
 
     return (
         <div className={style.wrapper}>
-            <SearchSection />
-            {!searchResultIsOpen && (
+            <Heading size={'small'} className={style.heading}>
+                {'Søk'}
+            </Heading>
+            <form
+                role={'search'}
+                onSubmit={(e) => {
+                    e.preventDefault();
+                    if (!searchQuery) {
+                        return;
+                    }
+
+                    searchData();
+                }}
+            >
+                <Search
+                    label={'Søk'}
+                    value={searchQuery}
+                    onChange={(value) => setSearchQuery(value)}
+                />
+            </form>
+            {searchResultIsOpen ? (
+                <SearchResult
+                    isLoading={isLoading}
+                    searchResult={searchResult}
+                    closeSearchResult={closeSearchResult}
+                />
+            ) : (
                 <Tabs defaultValue="no" onChange={(locale) => setSelectedLocale(locale as Locale)}>
                     <Tabs.List>
                         {locales.map((locale) => (
